# openapi/users.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/roles": {
      "get": {
        "operationId": "Role.listRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "OK"
                      ]
                    },
                    "msg": {
                      "type": "string",
                      "enum": [
                        ""
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Role"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "description": "Gửi kèm khi báo lỗi để tra log"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": {},
                      "description": "Luôn có requestId; endpoint list thêm nextCursor; khi lỗi thêm tham số của lỗi (issues, count, max…)"
                    }
                  },
                  "required": [
                    "success",
                    "code",
                    "msg",
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Danh sách role",
        "tags": [
          "Roles"
        ],
        "description": "**Quyền cần có:** `role:read`"
      },
      "post": {
        "operationId": "Role.createRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRole"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "OK"
                      ]
                    },
                    "msg": {
                      "type": "string",
                      "enum": [
                        ""
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Role"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "description": "Gửi kèm khi báo lỗi để tra log"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": {},
                      "description": "Luôn có requestId; endpoint list thêm nextCursor; khi lỗi thêm tham số của lỗi (issues, count, max…)"
                    }
                  },
                  "required": [
                    "success",
                    "code",
                    "msg",
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Tạo role",
        "tags": [
          "Roles"
        ],
        "description": "**Quyền cần có:** `role:create`"
      }
    },
    "/api/v1/admin/roles/{id}": {
      "put": {
        "operationId": "Role.updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRole"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "OK"
                      ]
                    },
                    "msg": {
                      "type": "string",
                      "enum": [
                        ""
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Role"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "description": "Gửi kèm khi báo lỗi để tra log"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": {},
                      "description": "Luôn có requestId; endpoint list thêm nextCursor; khi lỗi thêm tham số của lỗi (issues, count, max…)"
                    }
                  },
                  "required": [
                    "success",
                    "code",
                    "msg",
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Sửa role",
        "tags": [
          "Roles"
        ],
        "description": "**Quyền cần có:** `role:update`"
      },
      "delete": {
        "operationId": "Role.removeRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Xoá role",
        "tags": [
          "Roles"
        ],
        "description": "**Quyền cần có:** `role:delete`"
      }
    },
    "/api/v1/admin/users/{id}/roles": {
      "get": {
        "operationId": "Role.getUserRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "OK"
                      ]
                    },
                    "msg": {
                      "type": "string",
                      "enum": [
                        ""
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserRoles"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "description": "Gửi kèm khi báo lỗi để tra log"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": {},
                      "description": "Luôn có requestId; endpoint list thêm nextCursor; khi lỗi thêm tham số của lỗi (issues, count, max…)"
                    }
                  },
                  "required": [
                    "success",
                    "code",
                    "msg",
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Role hiện tại của một user",
        "tags": [
          "Roles"
        ],
        "description": "**Quyền cần có:** `role:read`"
      },
      "put": {
        "operationId": "Role.setUserRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUserRoles"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "OK"
                      ]
                    },
                    "msg": {
                      "type": "string",
                      "enum": [
                        ""
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserRoles"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "description": "Gửi kèm khi báo lỗi để tra log"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": {},
                      "description": "Luôn có requestId; endpoint list thêm nextCursor; khi lỗi thêm tham số của lỗi (issues, count, max…)"
                    }
                  },
                  "required": [
                    "success",
                    "code",
                    "msg",
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase": []
          }
        ],
        "summary": "Gán lại role cho user",
        "tags": [
          "Roles"
        ],
        "description": "**Quyền cần có:** `role:assign`"
      }
    },
...
  },
  "info": {
    "title": "C9 Map · User & Auth",
    "description": "Ngôn ngữ: header `Accept-Language: vi | en` (mặc định vi).",
    "version": "1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "supabase": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "code": {
            "type": "string",
            "description": "Mã lỗi SCREAMING_SNAKE, client rẽ nhánh theo mã này"
          },
          "msg": {
            "type": "string",
            "description": "Câu đã dịch theo header Accept-Language (vi mặc định, en)"
          },
          "data": {
            "type": "string",
            "nullable": true,
            "enum": [
              null
            ]
          },
          "meta": {
            "type": "object",
            "properties": {
              "requestId": {
                "type": "string",
                "description": "Gửi kèm khi báo lỗi để tra log"
              }
            },
            "required": [
              "requestId"
            ],
            "additionalProperties": {},
            "description": "Luôn có requestId; endpoint list thêm nextCursor; khi lỗi thêm tham số của lỗi (issues, count, max…)"
          }
        },
        "required": [
          "success",
          "code",
          "msg",
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
...
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "code": {
            "type": "string",
            "minLength": 2,
            "maxLength": 30,
            "pattern": "^[a-z][a-z0-9_]*$"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "isSystem": {
            "type": "boolean"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "description",
          "isSystem",
          "permissions"
        ],
        "additionalProperties": false
      },
      "CreateRole": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 2,
            "maxLength": 30,
            "pattern": "^[a-z][a-z0-9_]*$"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "permissions": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 60,
              "pattern": "^[a-z][a-z_]*:[a-z_]+$"
            }
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "UpdateRole": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "permissions": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 60,
              "pattern": "^[a-z][a-z_]*:[a-z_]+$"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "UserRoles": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 2,
              "maxLength": 30,
              "pattern": "^[a-z][a-z0-9_]*$"
            }
          }
        },
        "required": [
          "id",
          "roles"
        ],
        "additionalProperties": false
      },
      "SetUserRoles": {
        "type": "object",
        "properties": {
          "roles": {
            "minItems": 1,
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 2,
              "maxLength": 30,
              "pattern": "^[a-z][a-z0-9_]*$"
            }
          }
        },
        "required": [
          "roles"
        ]
      },
...
    }
  }
}